% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_one_setup.R
\name{analysis_one_setup}
\alias{analysis_one_setup}
\title{Title}
\usage{
analysis_one_setup(
  setup_number,
  files_folder,
  run_name,
  n_runs_per_setup,
  date
)
}
\arguments{
\item{date}{}
}
\description{
Title
}
