% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_offset_time.R
\name{as_offset_time}
\alias{as_offset_time}
\title{Convert Date to Elapsed Time}
\usage{
as_offset_time(x, start_date)
}
\arguments{
\item{x}{Date}

\item{start_date}{Initial timepoint, in decimal years}
}
\description{
Convert Date to Elapsed Time
}
\examples{
\dontrun{ as_offset_time(lubridate::mdy("2/10/2010"), 2010) }

}
\keyword{internal}
