% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_aft_runs_new_data.R
\name{batch_aft_runs_new_data}
\alias{batch_aft_runs_new_data}
\title{batch_aft_runs_new_data}
\usage{
batch_aft_runs_new_data(
  iterations = 10,
  covariate_effect_vector = c(0),
  covariate_list = NULL,
  covariate_names = NULL,
  n = 100,
  t_dist = function(n) {
     runif(n, min = 0, max = 1)
 },
  pi = function(t) {
     z <- 1 + 0 * t
     c(`1` = z)
 },
  `E[X|T,C]` = function(t, c) {
     case_when(c == "1" ~ 0 + 0 * t, TRUE ~ NaN)
 },
  sd_vector = c(`1` = 1),
  low_con = 2^-4,
  high_con = 2^4,
  type_list = c("loglogistic", "lognormal", "weibull", "gaussian", "exponential",
    "logistic"),
  time = "t",
  MIC_breakpoint = 1,
  stderr = FALSE
)
}
\arguments{
\item{stderr}{}
}
\description{
Run single_aft_run_new_data over a specified number of iterations.
For each iteration a new data set is generated and all six models are fitted to the data set.
}
