% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_fold_likelihood_all.R
\name{calculate_fold_likelihood_all}
\alias{calculate_fold_likelihood_all}
\title{Title}
\usage{
calculate_fold_likelihood_all(
  testing_set,
  trained_mu_model,
  trained_pi_model,
  approach = "full",
  fixed_side = NULL,
  extra_row = FALSE,
  ncomp = 2
)
}
\arguments{
\item{extra_row}{}
}
\description{
Title
}
\keyword{internal}
