% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_pi_area.R
\name{calculate_pi_area}
\alias{calculate_pi_area}
\title{Title}
\usage{
calculate_pi_area(
  results,
  settings,
  directionality,
  analysis,
  single_model_output
)
}
\arguments{
\item{single_model_output}{}
}
\description{
Title
}
