% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_single_covariate_effect.R
\name{calculate_single_covariate_effect}
\alias{calculate_single_covariate_effect}
\title{calculate_single_covariate_effect}
\usage{
calculate_single_covariate_effect(
  covariate_entry_column,
  covariate_list_of_lists
)
}
\arguments{
\item{covariate_list_of_lists}{}
}
\description{
used in the find_covariate_effects function on each covariate to find the individual covariate effect for each observation in simulated data
}
