% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture_error_measures_one_batch.R
\name{capture_error_measures_one_batch}
\alias{capture_error_measures_one_batch}
\title{Title}
\usage{
capture_error_measures_one_batch(
  location,
  format,
  array_name,
  date,
  i,
  number_of_batches
)
}
\arguments{
\item{number_of_batches}{}
}
\description{
Title
}
