% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture_error_measures_one_run.R
\name{capture_error_measures_one_run}
\alias{capture_error_measures_one_run}
\title{Title}
\usage{
capture_error_measures_one_run(
  individual_run,
  intercepts,
  trends,
  sigma,
  pi_int,
  pi_trend,
  sigma_tolerance = c(0.05, 100),
  pi_int_tolerance = c(1e-04, 0.9999),
  pi_trend_tolerance = c(-4, 4),
  intercepts_tolerance = 100,
  trends_tolerance = 100
)
}
\arguments{
\item{trends_tolerance}{}
}
\description{
Title
}
