% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture_error_measures_one_run.R
\name{capture_error_measures_one_run_both_directions_safety}
\alias{capture_error_measures_one_run_both_directions_safety}
\title{Title}
\usage{
capture_error_measures_one_run_both_directions_safety(
  individual_run,
  intercepts,
  trends,
  sigma,
  pi,
  sigma_tolerance = c(0.05, 100),
  pi_tolerance = c(0.05, 0.95),
  intercepts_tolerance = 100,
  trends_tolerance = 100
)
}
\arguments{
\item{trends_tolerance}{}
}
\description{
Title
}
