% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censoring_post_info.R
\name{censoring_post_info}
\alias{censoring_post_info}
\title{Title}
\usage{
censoring_post_info(possible_data, settings, comparison = "model_weighted")
}
\arguments{
\item{comparison}{}
}
\description{
Title
}
