% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_array_complete.R
\name{check_array_complete}
\alias{check_array_complete}
\title{check_array_complete}
\usage{
check_array_complete(number_of_batches, format, location, array_name, date)
}
\arguments{
\item{date}{}
}
\description{
check_array_complete
}
