% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_for_excessive_censoring.R
\name{check_for_excessive_censoring}
\alias{check_for_excessive_censoring}
\title{Title}
\usage{
check_for_excessive_censoring(single_model_output, cutoff = 0.9)
}
\arguments{
\item{cutoff}{}
}
\description{
Title
}
