% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_survreg_iteration_maxout.R
\name{check_survreg_iteration_maxout}
\alias{check_survreg_iteration_maxout}
\title{Check Survreg Iteration Maxout}
\usage{
check_survreg_iteration_maxout(mu_models, ncomp, maxiter_survreg)
}
\arguments{
\item{mu_models}{list of survreg objects fitted to each component by the EM algorithm}

\item{ncomp}{number of components being fitted by the EM algorithm}

\item{maxiter_survreg}{maximum number of iterations for survreg to fit the model}
}
\description{
Returns a logical value if the number of iterations for survreg to converge is equal to the maximum number of iterations specified
}
\keyword{internal}
