% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_weigh_prop_cens.R
\name{check_weigh_prop_cens}
\alias{check_weigh_prop_cens}
\title{Title}
\usage{
check_weigh_prop_cens(df = a, comp_i = "2", side = "rc", cutoff = 0.9)
}
\arguments{
\item{cutoff}{}
}
\description{
Get a weighted sum of the proportion of censored observations in each component from possible data
}
