% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component_mean_alt.R
\name{component_mean_alt}
\alias{component_mean_alt}
\title{component_mean - alternate version}
\usage{
component_mean_alt(
  n = 100,
  t_dist = function(n) {     runif(n, min = 0, max = 1) },
  pi = function(t) {     z <- 0.5 + 0.2 * t     c(`1` = z, `2` = 1 - z) },
  `E[X|T,C]` = function(t, c) {     case_when(c == "1" ~ 3 + t + 2 * t^2 - sqrt(t), c
    == "2" ~ 3 * t, TRUE ~ NA) }
)
}
\arguments{
\item{`E[X|T, C]`}{}
}
\value{

}
\description{
component_mean - alternate version
}
