% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_effect_total.R
\name{covariate_effect_total}
\alias{covariate_effect_total}
\title{covariate_effect_total}
\usage{
covariate_effect_total(data_tibble, covariate_effect_vector)
}
\arguments{
\item{covariate_effect_vector}{}
}
\description{
uses model.matrix to
}
\keyword{internal}
