% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_year.R
\name{create_year}
\alias{create_year}
\title{create_year}
\usage{
create_year(
  sample_size_dist = "normal",
  nyears = 5,
  norm_mean = 50,
  norm_sd = 10,
  unif_min,
  unif_max
)
}
\arguments{
\item{unif_max}{}
}
\value{

}
\description{
create_year
}
