% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_data_set.R
\name{describe_data_set}
\alias{describe_data_set}
\title{describe_data_set}
\usage{
describe_data_set(
  i,
  n,
  intercepts,
  trends,
  sigma,
  pi,
  nyears,
  converge_incorrectly_vector,
  failure_to_converge_vector,
  scale = "log"
)
}
\arguments{
\item{scale}{}
}
\description{
describe_data_set
}
