% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_covariates.R
\name{draw_covariates}
\alias{draw_covariates}
\title{draw_covariates}
\usage{
draw_covariates(input, cov_list)
}
\arguments{
\item{cov_list}{}
}
\description{
draw_covariates function uses the draw_categorical_covariate and draw_numerical_covariate to add covariates to a dataframe or tibble of drawn observations
}
\details{
the cov_list is a list object of vectors with a vector for each covariate to be added, either numeric or categorical
the first value in each vector should be either "numeric" or "categorical"

for numeric variables, the second value should specify whether it is drawn from a normal or uniform distribution using "normal" or "uniform"
the next two values are the mean and sd if drawn from a normal distribution or min and max if drawn from a uniform distribution

for categorical variables, the remaining values in the vector will be probabilities for each level of the variable, which will be labeled as a, b, c, and so on

example:
cov_list <- list(
c("numeric", "normal", 40, 4),
c("categorical", 0.3, 0.4, 0.3),
c("categorical", 0.5, 0.2, 0.3),
c("categorical", 0.2, 0.2, 0.2, 0.2, 0.2),
c("numeric", "uniform", 1, 10))
}
\keyword{internal}
