% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_observations.R
\name{draw_observations}
\alias{draw_observations}
\title{draw_observations}
\usage{
draw_observations(
  year = 0,
  initial_lower_mean = -1,
  initial_upper_mean = 1,
  initial_lower_std_dev = 1,
  initial_upper_std_dev = 1,
  initial_lower_slp = 0,
  initial_upper_slp = 0,
  initial_lower_slp_qd = 0,
  initial_upper_slp_qd = 0,
  initial_lower_lambda = 0.5,
  initial_upper_lambda = 0.5,
  initial_lower_slp_lambda = 0,
  initial_upper_slp_lambda = 0
)
}
\arguments{
\item{initial_upper_slp_lambda}{}
}
\description{
draws a single observation from a two-component gaussian mixture model
}
