% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_algorithm.R
\name{EM_algorithm}
\alias{EM_algorithm}
\title{Fit a Full Mixture Model Using the EM Algorithm}
\usage{
EM_algorithm(
  visible_data,
  model = "surv",
  mu_formula = Surv(time = left_bound, time2 = right_bound, type = "interval2") ~
    pspline(t, df = 4),
  pi_formula = c == "2" ~ s(t),
  max_it = 3000,
  ncomp = 2,
  tol_ll = 1e-06,
  browse_at_end = FALSE,
  browse_each_step = FALSE,
  plot_visuals = FALSE,
  prior_step_plot = FALSE,
  pause_on_likelihood_drop = FALSE,
  pi_link = "logit",
  verbose = 3,
  model_coefficient_tolerance = 1e-05,
  maxiter_survreg = 30,
  initial_weighting = 1,
  sd_initial = 0.2,
  stop_on_likelihood_drop = FALSE,
  n_models = 100,
  seed = NULL,
  randomize = "all",
  non_linear_term = "t",
  covariates = NULL,
  scale = NULL
)
}
\arguments{
\item{visible_data}{Data frame, data including the left and right bound of the MICs (use import_mics_with_metadata to format correctly) and any covariates (including the non-linear term)}

\item{model}{String, "surv" (equivalently "pspline") or "polynomial" to fit the mu models with survreg(), or "mgcv" to fit them with mgcv::gam(). Must match the terms used in mu_formula.}

\item{mu_formula}{A formula for a survreg object from the survival package, left side of equation should be a surv object using "interval2" format, right side should be the non-linear term (polynomial or pspline) and any covariates. Can be a single formula or a list of formulas where length is equal to the number of components where the trend in the mean is being estimated.}

\item{pi_formula}{Formula, for the component weight model. Model is fit using mgcv's gam function. Nonlinear terms include s() and lo(). Basis of s() function can be changed using bs argument to s(). Use c == "2" for the left side of the formula.}

\item{max_it}{Numeric, maximum number of iterations for the EM algorithm for any given model fitting.}

\item{ncomp}{Numeric, number of components to be fitted. When fitting a reduced model where one component is not estimated, that component should still contribute to the value in ncomp. E.g. a reduced model where the upper component is fixed and mu for the lower component is being estimated has a value of ncomp = 2.}

\item{tol_ll}{Numeric, maximum tolerance for change in likelihood between steps of the algorithm for model convergence to be achieved.}

\item{browse_at_end}{For internal model testing (currently unused)}

\item{browse_each_step}{For internal model testing, logical, if TRUE calls browser() at the end of each EM step}

\item{plot_visuals}{For internal model testing, logical, if TRUE (together with browse_each_step) plots the fit at each step}

\item{prior_step_plot}{For internal model testing, logical, if TRUE the step plots also show the fit from the previous step}

\item{pause_on_likelihood_drop}{For internal model testing, logical, if TRUE calls browser() when the log likelihood decreases between steps}

\item{pi_link}{String: "logit" or "identity", link function for the generalized linear model fit for the pi model (component weights).}

\item{verbose}{Numeric, controls amount of information printed during model fitting}

\item{model_coefficient_tolerance}{Numeric, maximum tolerance for change in model coefficients (including spline terms) between steps of the algorithm for model convergence to be achieved.}

\item{maxiter_survreg}{Maximum iterations used in survreg model fitting, default is 30.}

\item{initial_weighting}{Numeric, if 1: initial observation weights are estimated using linear regression at the highest and lowest tested concentrations. If 2, used a randomized start suitable for simulation studies on model validity but otherwise not recommended. If 3 or greater, fits a linear models to the components and estimates initial weights based on this model fit.}

\item{sd_initial}{Numeric, value greater than 0 and less than 1. Proportion of the range from the highest concentration to lowest concentration that is used as the initial estimate of sigma for the estimated components. Default is 0.2}

\item{stop_on_likelihood_drop}{For internal model testing, logical, if TRUE stops the algorithm (with converge = "likelihood decreased") when the log likelihood decreases between steps}

\item{n_models}{Numeric, only used for the randomized start (initial_weighting = 2): the random perturbation of the initial component means has standard deviation sd_initial * (range of tested concentrations) / sqrt(n_models)}

\item{seed}{Numeric or NULL, if not NULL the random seed is set to this value before fitting}

\item{randomize}{String, only used for the randomized start (initial_weighting = 2): which starting values to randomize, one of "all", "mu", "sigma", or "pi"}

\item{non_linear_term}{String, non-linear term to be included in the model. Variable in the pspline term in the pspline model or in the polynomial term in the polynomial model.}

\item{covariates}{String, covariates to be included in mu model aside from the non-linear term.}

\item{scale}{String or NULL, the scale of left_bound and right_bound in visible_data: "log" (log2 MIC; "log2" and "fold" are also accepted) or "MIC" (concentration). Only used when visible_data does not already have a "scale" attribute (import_mics_with_metadata() and simulate_mics() set one); data on the "MIC" scale are converted to log2 before fitting.}
}
\value{
If ncomp = 1, a list with elements possible_data, mu_model, converge,
  ncomp, and likelihood. Otherwise a list with elements
  \describe{
    \item{likelihood}{tibble of the log likelihood (and related checks) at each step}
    \item{model}{the value of model used}
    \item{possible_data}{the data with one row per observation and component,
      including the posterior component probabilities \code{P(C=c|y,t)}}
    \item{pi_model}{the fitted \code{gam} model for the component weights}
    \item{mu_model}{list of the fitted models for the component means}
    \item{steps}{number of EM steps run}
    \item{converge}{"YES", "NO" (a model fit failed), "iterations" (max_it reached),
      or "likelihood decreased"}
    \item{ncomp}{number of components}
    \item{prior_step_models}{the models and log likelihood from the previous step}
    \item{seed, random_start_set, sd_initial, mu_formula}{settings used for the fit}
  }
}
\description{
Fits a mixture of Gaussian components to interval-censored log2 MIC data
using the EM algorithm, estimating a model for the mean (mu) of every
component and a generalized additive model for the component weights (pi).
This is the fitting routine that fit_EM() calls when approach = "full";
most users should call fit_EM() instead.
}
\examples{
\donttest{
data = simulate_mics()
fit = EM_algorithm(data, max_it = 300, verbose = 0)
fit$converge
}
}
