% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_fm_surv_single_run.R
\name{EM_fm_surv_single_run}
\alias{EM_fm_surv_single_run}
\title{Title}
\usage{
EM_fm_surv_single_run(
  seed,
  iter,
  visible_data,
  mu_formula,
  pi_formula,
  max_it,
  ncomp,
  tol_ll,
  pi_link,
  verbose,
  initial_weighting,
  model_coefficient_tolerance,
  maxiter_survreg,
  sd_initial,
  randomize,
  n_models
)
}
\arguments{
\item{sd_initial}{}
}
\description{
Title
}
