% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_measures_one_batch.R
\name{error_measures_one_batch}
\alias{error_measures_one_batch}
\title{error_measures_one_batch}
\usage{
error_measures_one_batch(
  location,
  format,
  array_name,
  date,
  i,
  intercepts,
  trends,
  sigma,
  pi,
  sigma_tolerance = c(0.05, 100),
  pi_tolerance = c(0.05, 0.95),
  intercepts_tolerance = 100,
  trends_tolerance = 100
)
}
\arguments{
\item{trends_tolerance}{}
}
\description{
error_measures_one_batch
}
