% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_measures_one_run.R
\name{error_measures_one_run}
\alias{error_measures_one_run}
\title{error_measures_one_run}
\usage{
error_measures_one_run(individual_run, intercepts, trends, sigma)
}
\arguments{
\item{sigma}{}
}
\description{
Extracts error measures from a single simulation. Useful when used in a map function applied to a batch of simulations.
}
