% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_measures_one_run_both_directions.R
\name{error_measures_one_run_both_directions}
\alias{error_measures_one_run_both_directions}
\title{Title}
\usage{
error_measures_one_run_both_directions(
  individual_run,
  intercepts,
  trends,
  sigma,
  pi,
  error_threshold = 10
)
}
\arguments{
\item{error_threshold}{}
}
\description{
Title
}
