% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_covariate_effects.R
\name{find_covariate_effects}
\alias{find_covariate_effects}
\title{find_covariate_effects}
\usage{
find_covariate_effects(input, cov_list)
}
\arguments{
\item{input}{simulated data goes here, first two columns of that should be year and base data}

\item{cov_list}{this is the covariate list where each entry is a list of 2 vectors, the first describes the distribution of the variable and the second describes its effects}
}
\description{
uses map from purrr package to find the single covariate effect over each covariate
}
