% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_epsilon.R
\name{find_epsilon}
\alias{find_epsilon}
\title{find_epsilon}
\usage{
find_epsilon(
  year,
  sd_1,
  sd_2,
  mean_1_trend,
  mean_2_trend,
  mean_1_intercept,
  mean_2_intercept,
  pi_1_trend,
  pi_1_intercept
)
}
\arguments{
\item{pi_1_intercept}{}
}
\value{

}
\description{
finds the base value for each observation by drawing from a gaussian mixture with 2 components
}
