% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_E_step.R
\name{first_E_step}
\alias{first_E_step}
\title{Title}
\usage{
first_E_step(
  initial_weighting,
  visible_data,
  plot_visuals,
  sd_initial = 0.2,
  ncomp = 2,
  randomize = "all",
  n_models = 100,
  model,
  non_linear_term,
  covariates,
  pi_formula,
  max_it,
  tol_ll,
  pi_link,
  model_coefficient_tolerance
)
}
\arguments{
\item{model_coefficient_tolerance}{}
}
\description{
Title
}
\keyword{internal}
