% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_E_step_reduced.R
\name{first_E_step_reduced}
\alias{first_E_step_reduced}
\title{Title}
\usage{
first_E_step_reduced(
  initial_weighting,
  visible_data,
  plot_visuals,
  sd_initial = 0.2,
  ncomp = 2,
  non_linear_term,
  covariates,
  pi_formula,
  max_it,
  tol_ll,
  pi_link,
  model_coefficient_tolerance,
  fixed_side,
  extra_row,
  ECOFF,
  model,
  verbose
)
}
\arguments{
\item{ECOFF}{Numeric, highest value of the WT distribution on the log2 scale}

\item{model}{}
}
\description{
Title
}
\keyword{internal}
