% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_aft.R
\name{fit_aft}
\alias{fit_aft}
\title{fit_aft}
\usage{
fit_aft(
  observed_values,
  covariate_data_frame,
  type,
  low_con = 2^-4,
  high_con = 2^4,
  tested_concentrations = log2(low_con):log2(high_con),
  summary = FALSE
)
}
\arguments{
\item{summary}{}
}
\value{

}
\description{
fit_aft
}
