% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_all_mu_models.surv.R
\name{fit_all_mu_models.surv}
\alias{fit_all_mu_models.surv}
\title{Fit All Mu Models}
\usage{
fit_all_mu_models.surv(possible_data, ncomp, mu_formula, maxiter_survreg)
}
\arguments{
\item{possible_data}{data frame iterated on by the EM algorithm, should contain left_bound, right_bound, c, P(C=c|y,t), and any covariates, such as time, e.g. "t"}

\item{ncomp}{number of components being fitted by the EM algorithm}

\item{mu_formula}{formula being used to fit the model for mu, should contain a Surv object for an outcome and any covariates of interest, accepts pspline and other arguments to survreg}

\item{maxiter_survreg}{maximum number of iterations for survreg to fit the model}
}
\description{
Loops the function that fits survreg to each component over all the iterations (1 through the number of components being fitted by the EM algorithm)
}
