% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lr.R
\name{fit_lr}
\alias{fit_lr}
\title{fit_lr}
\usage{
fit_lr(df, MIC_breakpoint, summary = FALSE)
}
\arguments{
\item{df}{data frame of year, observed values (colname is observed_value), and covariates}

\item{MIC_breakpoint}{on MIC scale, breakpoint for S vs I}

\item{summary}{whether the output is the summary.glm  (TRUE) or the glm object}
}
\value{

}
\description{
fit_lr
}
