% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model_safety.R
\name{fit_model_safety}
\alias{fit_model_safety}
\title{Title}
\usage{
fit_model_safety(
  visible_data = prep_sim_data_for_em(),
  formula = Surv(time = left_bound, time2 = right_bound, type = "interval2") ~ t,
  max_it = 3000,
  ncomp = 2,
  tol_ll = 1e-06,
  browse_at_end = FALSE,
  browse_each_step = FALSE,
  plot_visuals = FALSE,
  verbose = 3,
  low_con = 2^-4,
  high_con = 2^4
)
}
\arguments{
\item{high_con}{}
}
\description{
Title
}
