% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model_safety_pi.R
\name{fit_model_safety_pi}
\alias{fit_model_safety_pi}
\title{Title}
\usage{
fit_model_safety_pi(
  visible_data = prep_sim_data_for_em(),
  formula = Surv(time = left_bound, time2 = right_bound, type = "interval2") ~ pspline(t,
    df = 0, calc = TRUE),
  formula2 = c == "2" ~ s(t),
  max_it = 3000,
  ncomp = 2,
  tol_ll = 1e-06,
  browse_at_end = FALSE,
  browse_each_step = FALSE,
  plot_visuals = FALSE,
  pi_link = "logit",
  verbose = 3,
  maxiter_survreg = 30
)
}
\arguments{
\item{maxiter_survreg}{}
}
\description{
Title
}
