% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mu_model.R
\name{fit_mu_model}
\alias{fit_mu_model}
\title{Fit Mu Model}
\usage{
fit_mu_model(possible_data, pred_comp, mu_formula, maxiter_survreg = 30)
}
\arguments{
\item{possible_data}{data frame iterated on by the EM algorithm, should contain left_bound, right_bound, c, P(C=c|y,t), and any covariates, such as time, e.g. "t"}

\item{pred_comp}{numeric variable, which component is this model being fitted to, e.g. 2}

\item{mu_formula}{formula being used to fit the model for mu, should contain a Surv object for an outcome and any covariates of interest, accepts pspline and other arguments to survreg}

\item{maxiter_survreg}{maximum number of iterations for survreg to fit the model}
}
\description{
Function to filter the data to the desired component using the variable c and fit survreg to the data using a formula supplied by the user, typically passed in through the EM algorithm
}
\keyword{internal}
