% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_pi_model.R
\name{fit_pi_model}
\alias{fit_pi_model}
\title{Fit Pi Model}
\usage{
fit_pi_model(pi_formula, pi_link, possible_data)
}
\arguments{
\item{pi_formula}{formula used to fit the binomial model for pi, the component weights. The left hand side should designate the level of c, component to classify as an event, and the right hand side should include the covariates. Accepts smoothing splines s() and local regression lo() from the gam package. Example of formula: c == "2" ~ s(t)}

\item{pi_link}{string specifying whether to use a logit or identity link function. Possible options are "logit" and "identity"}

\item{possible_data}{data frame iterated on by the EM algorithm, should contain left_bound, right_bound, c, P(C=c|y,t), and any covariates, such as time, e.g. "t"}
}
\description{
Fits a weighted binomial model using either a logit or identity function to estimate covariate effects on component weights, pi. Uses the gam function from the gam package.
}
