% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_single_component_model_surv.R
\name{fit_single_component_model_surv}
\alias{fit_single_component_model_surv}
\title{Title}
\usage{
fit_single_component_model_surv(
  visible_data,
  mu_formula,
  maxiter_survreg,
  verbose
)
}
\arguments{
\item{verbose}{}
}
\description{
Title
}
