% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_spaft.R
\name{fit_spaft}
\alias{fit_spaft}
\title{fit_spaft}
\usage{
fit_spaft(
  observed_values,
  covariate_data_frame,
  low_con = 2^-4,
  high_con = 2^4,
  tested_concentrations = log2(low_con):log2(high_con)
)
}
\arguments{
\item{tested_concentrations}{}
}
\value{

}
\description{
apply semiparametric accelerated failure time model using smoothSurv
}
