% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_sim_in_1_function.R
\name{full_sim_in_1_function}
\alias{full_sim_in_1_function}
\title{full_sim_in_1_function}
\usage{
full_sim_in_1_function(
  i,
  n = 100,
  t_dist = function(n) {
     runif(n, min = 0, max = 1)
 },
  pi = function(t) {
     z <- 0.5 + 0.2 * t
     c(`1` = z, `2` = 1 - z)
 },
  `E[X|T,C]` = function(t, c) {
     case_when(c == "1" ~ 3 + t + 2 * t^2 - sqrt(t), c ==
    "2" ~ 3 * t, TRUE ~ NaN)
 },
  sd_vector = c(`1` = 1, `2` = 2),
  covariate_list,
  covariate_effect_vector,
  low_con = 2^-4,
  high_con = 2^4,
  scale = "log",
  formula = Surv(time = left_bound, time2 = right_bound, type = "interval2") ~ 0 + c +
    strata(c) + t:c,
  max_it = 3000,
  ncomp = 2,
  tol_ll = 1e-06,
  silent = FALSE,
  ...
)
}
\arguments{
\item{tol_ll}{}
}
\description{
full_sim_in_1_function
}
