% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_path_sim.R
\name{gen_path_sim}
\alias{gen_path_sim}
\title{gen_path_sim}
\usage{
gen_path_sim(location, format, array_name, date, i)
}
\arguments{
\item{i}{}
}
\description{
Creates names of individual batches to load in, useful in combination with a map function and loadRData
}
