% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_sim_parameters.R
\name{grab_sim_parameters}
\alias{grab_sim_parameters}
\title{grab_sim_parameters}
\usage{
grab_sim_parameters(parameter_log, array_name, date, id_col)
}
\arguments{
\item{id_col}{}
}
\description{
from simulation log
}
