% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_mics.R
\name{import_mics}
\alias{import_mics}
\title{import_mics}
\usage{
import_mics(
  mic_column,
  code_column = NULL,
  combination_agent = FALSE,
  log_reg_value = FALSE
)
}
\arguments{
\item{combination_agent}{}
}
\description{
import_mics
}
