% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_weighting_fit_linear_model.R
\name{initial_weighting_fit_linear_model}
\alias{initial_weighting_fit_linear_model}
\title{Title}
\usage{
initial_weighting_fit_linear_model(
  visible_data,
  non_linear_term,
  covariates,
  pi_formula,
  max_it,
  ncomp,
  tol_ll,
  pi_link,
  verbose,
  model_coefficient_tolerance,
  sd_initial
)
}
\arguments{
\item{sd_initial}{}
}
\description{
Title
}
\keyword{internal}
