% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_weighting_safety.R
\name{initial_weighting_safety}
\alias{initial_weighting_safety}
\title{Title}
\usage{
initial_weighting_safety(visible_data, censored_side, extra_row)
}
\arguments{
\item{extra_row}{}
}
\description{
Title
}
