% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternative_initial_weighting.R
\name{initial_weighting_staggered_weighting_by_distance_from_median_and_boundary}
\alias{initial_weighting_staggered_weighting_by_distance_from_median_and_boundary}
\title{Title}
\usage{
initial_weighting_staggered_weighting_by_distance_from_median_and_boundary(
  visible_data
)
}
\arguments{
\item{visible_data}{}
}
\description{
Title
}
