% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadRData.R
\name{loadRData}
\alias{loadRData}
\title{loadRData}
\usage{
loadRData(fileName)
}
\arguments{
\item{fileName}{}
}
\description{
Borrowed from https://stackoverflow.com/a/25455968 to import Rdata files and assign them to a variable name
}
