% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_full_run_function.R
\name{local_full_run_function}
\alias{local_full_run_function}
\title{local_full_run_function}
\usage{
local_full_run_function(
  args,
  batch_size = 10,
  run_name,
  n,
  t_dist,
  pi,
  `E[X|T,C]`,
  sd_vector,
  covariate_list = NULL,
  covariate_effect_vector = c(0),
  covariate_names = NULL,
  low_con,
  high_con,
  scale = "log",
  formula,
  formula2,
  max_it = 3000,
  ncomp = 2,
  tol_ll = 1e-06,
  maxiter_survreg = 30,
  pi_function,
  pi_link,
  verbose = 3,
  allow_safety = TRUE,
  cutoff = 0.9,
  fms_only,
  initial_weighting = 1
)
}
\arguments{
\item{cutoff}{}
}
\description{
local_full_run_function
}
