% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mic.sim-package.R
\docType{package}
\name{mic.sim-package}
\alias{mic.sim}
\alias{mic.sim-package}
\title{mic.sim: Simulating and Fitting Models to MIC Data}
\description{
Analyze minimum inhibitory concentration (MIC) data: import or simulate data, view the distribution, fit models, and visualize the fits. Models are latent Gaussian mixtures for interval-censored log2 MIC values, in which the mixture weights and component means vary smoothly over time, to distinguish and track the wild-type and non-wild-type subpopulations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ajmichaelucd/mic.sim}
  \item \url{https://ajmichaelucd.github.io/mic.sim/}
  \item Report bugs at \url{https://github.com/ajmichaelucd/mic.sim/issues}
}

}
\author{
\strong{Maintainer}: Alec Michael \email{alec.j.michael@gmail.com} [copyright holder]

}
\keyword{internal}
