% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixcensoredInt.R
\name{mixcensoredInt}
\alias{mixcensoredInt}
\title{Title}
\usage{
mixcensoredInt(
  y1,
  y2,
  d,
  wt = rep(1, length(y1)),
  dist = "gaussian",
  n,
  cluster = NULL,
  classify = "EM",
  maxiter = 10000,
  tol = 1e-06
)
}
\arguments{
\item{tol}{}
}
\description{
Title
}
