% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offset_time_as_date.R
\name{offset_time_as_date}
\alias{offset_time_as_date}
\title{Convert Elapsed Time to Date}
\usage{
offset_time_as_date(t, start_date)
}
\arguments{
\item{t}{time in decimal years since start_date}

\item{start_date}{start date in decimal years}
}
\description{
Convert Elapsed Time to Date
}
\examples{
\dontrun{ offset_time_as_date(2.109589, 2007) }
}
\keyword{internal}
