% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fm_step.R
\name{plot_fm_step}
\alias{plot_fm_step}
\title{Title}
\usage{
plot_fm_step(
  pi_model_new,
  mu_models_new,
  ncomp,
  possible_data,
  prior_step_plot = FALSE,
  i,
  mu_models_old = NULL,
  pi_model_old = NULL
)
}
\arguments{
\item{pi_model_old}{}
}
\description{
Title
}
