% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fms.R
\name{plot_fms}
\alias{plot_fms}
\title{Title}
\usage{
plot_fms(
  output,
  title,
  cens_dir,
  add_log_reg = FALSE,
  s_breakpoint = NULL,
  r_breakpoint = NULL
)
}
\arguments{
\item{title}{}
}
\description{
Title
}
