% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mixed_normal_density.R
\name{plot_mixed_normal_density}
\alias{plot_mixed_normal_density}
\title{plot_mixed_normal_density}
\usage{
plot_mixed_normal_density(
  mean = c(-2, 1),
  sigma = c(1, 0.8),
  pi = c(0.4, 0.6),
  min = -5,
  max = 5,
  low_con = -3,
  high_con = 3
)
}
\arguments{
\item{high_con}{}
}
\description{
plot_mixed_normal_density
}
