% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_start.R
\name{random_start}
\alias{random_start}
\title{Title}
\usage{
random_start(visible_data, ncomp, sd_parameter = 0.2, n_models, randomize)
}
\arguments{
\item{randomize}{either "all" (default), "mu", "sigma", "pi"}
}
\description{
Title
}
\keyword{internal}
