% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_start_safety.R
\name{random_start_safety}
\alias{random_start_safety}
\title{Title}
\usage{
random_start_safety(
  visible_data,
  censored_side,
  extra_row,
  randomize,
  sd_parameter,
  n_models
)
}
\arguments{
\item{n_models}{}
}
\description{
Title
}
