% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recreate_and_plot_data.R
\name{recreate_and_plot_data}
\alias{recreate_and_plot_data}
\title{recreate_and_plot_data}
\usage{
recreate_and_plot_data(
  set_numbers,
  n,
  intercepts,
  trends,
  sigma,
  pi,
  nyears,
  covariate_list,
  covariate_effect_vector,
  low_con,
  high_con,
  scale,
  converge_incorrectly_vector,
  failure_to_converge_vector
)
}
\arguments{
\item{failure_to_converge_vector}{}
}
\description{
recreate_and_plot_data
}
