% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recreate_data_set.R
\name{recreate_data_set}
\alias{recreate_data_set}
\title{recreate_data_set}
\usage{
recreate_data_set(i, n, intercepts, trends, sigma, pi, nyears)
}
\arguments{
\item{nyears}{}
}
\description{
recreate_data_set
}
