% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_failure_types.R
\name{report_failure_types}
\alias{report_failure_types}
\title{report_failure_types}
\usage{
report_failure_types(array_results)
}
\arguments{
\item{array_results}{}
}
\description{
report_failure_types
}
