% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{sim_data}
\alias{sim_data}
\title{sim_data}
\usage{
sim_data(
  nyears = 5,
  sample_size_dist = "normal",
  norm_mean = 100,
  norm_sd = 10,
  unif_min,
  unif_max
)
}
\arguments{
\item{unif_max}{}
}
\value{

}
\description{
Function to draw values for each year
}
