% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data_and_add_covariates.R
\name{sim_data_and_add_covariates}
\alias{sim_data_and_add_covariates}
\title{Title}
\usage{
sim_data_and_add_covariates(
  nyears = 5,
  sample_size_dist = "normal",
  norm_mean = 100,
  norm_sd = 10,
  unif_min,
  unif_max,
  covariate_list
)
}
\arguments{
\item{covariate_list}{}
}
\value{

}
\description{
Title
}
