% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pi_survreg_boot.R
\name{sim_pi_survreg_boot}
\alias{sim_pi_survreg_boot}
\title{Title}
\usage{
sim_pi_survreg_boot(df, fit, alpha = 0.05, nSims = 10000)
}
\arguments{
\item{nSims}{}
}
\description{
Title
}
\keyword{internal}
