% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pi_survreg_boot_mgcv.R
\name{sim_pi_survreg_boot_mgcv}
\alias{sim_pi_survreg_boot_mgcv}
\title{Title}
\usage{
sim_pi_survreg_boot_mgcv(
  df,
  fit,
  alpha = 0.05,
  nSims = 10000,
  distr = "gaussian"
)
}
\arguments{
\item{distr}{}
}
\description{
Title
}
