% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_mics_2.R
\name{simulate_mics_2}
\alias{simulate_mics_2}
\title{Title}
\usage{
simulate_mics_2(
  n = 100,
  t_dist = function(n) {
     runif(n, min = 0, max = 1)
 },
  pi = function(t) {
     z <- 0.5 + 0.2 * t
     c(`1` = z, `2` = 1 - z)
 },
  `E[X|T,C]` = function(t, c) {
     case_when(c == "1" ~ -1 + t, c == "2" ~ 1 + 0.3 * t,
    TRUE ~ NaN)
 },
  sd_vector = c(`1` = 1, `2` = 0.7),
  covariate_list = list(c("numeric", "normal", 0, 1), c("categorical", c(0.3, 0.4))),
  covariate_effect_vector = c(0, 0.2, -1),
  conc_limits_table = as_tibble(rbind(c("a", 2^-3, 2^3), c("b", 2^-4, 2^4)), .name_repair
    = "unique") \%>\% rename(covariate_2 = 1, low_cons = 2, high_cons = 3),
  low_con = 2^-4,
  high_con = 2^4,
  scale = "log"
)
}
\arguments{
\item{scale}{}
}
\description{
Title
}
