% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_1_set.R
\name{simulation_1_set}
\alias{simulation_1_set}
\title{Title}
\usage{
simulation_1_set(
  mic.seed,
  random_start_seeds,
  n,
  t_dist,
  pi,
  `E[X|T,C]`,
  sd_vector,
  low_con,
  high_con,
  scale = "log",
  model = "fm",
  mu_formula = Surv(time = left_bound, time2 = right_bound, type = "interval2") ~
    pspline(t, df = 0, caic = TRUE),
  pi_formula = c == "2" ~ s(t),
  max_it = 500,
  ncomp = 2,
  tol_ll = 1e-05,
  pi_link = "logit",
  verbose = 3,
  initial_weighting = 7,
  model_coefficient_tolerance = 1e-04,
  maxiter_survreg = 30,
  sd_initial = 0.2,
  randomize = "all",
  censored_side = NULL,
  extra_row = NULL
)
}
\arguments{
\item{extra_row}{}
}
\description{
Title
}
