% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_1_set_polynomial.R
\name{simulation_1_set_polynomial}
\alias{simulation_1_set_polynomial}
\title{Title}
\usage{
simulation_1_set_polynomial(
  mic.seed,
  n,
  t_dist,
  pi,
  `E[X|T,C]`,
  sd_vector,
  low_con,
  high_con,
  scale = "log",
  max_degree = 8,
  nfolds = 10,
  non_linear_term = "t",
  covariates = NULL,
  pi_formula = c == "2" ~ s(t),
  max_it = 500,
  ncomp = 2,
  tol_ll = 1e-05,
  pi_link = "logit",
  verbose = 3,
  initial_weighting = 7,
  model_coefficient_tolerance = 1e-04,
  maxiter_survreg = 30,
  sd_initial = 0.2,
  randomize = "all",
  censored_side = NULL,
  extra_row = NULL
)
}
\arguments{
\item{extra_row}{}
}
\description{
Title
}
