% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_aft_run_new_data.R
\name{single_aft_run_new_data}
\alias{single_aft_run_new_data}
\title{single_aft_run_new_data}
\usage{
single_aft_run_new_data(
  covariate_effect_vector = c(0),
  covariate_list = NULL,
  covariate_names = NULL,
  n = 100,
  t_dist = function(n) {     runif(n, min = 0, max = 1) },
  pi = function(t) {     z <- 1 + 0 * t     c(`1` = z) },
  complist = list(`1` = function(t) {     0 + 0 * t }),
  sd_vector = c(`1` = 1),
  low_con = 2^-4,
  high_con = 2^4,
  type_list = c("loglogistic", "lognormal", "weibull", "gaussian", "exponential",
    "logistic"),
  time = "t",
  MIC_breakpoint = 1,
  iteration = 1,
  stderr = FALSE
)
}
\arguments{
\item{stderr}{}
}
\value{

}
\description{
Generates a single data set and fits aft models to it, for each type list in type_list parameter
}
