% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statistics_one_run.R
\name{summary_statistics_one_run}
\alias{summary_statistics_one_run}
\title{Title}
\usage{
summary_statistics_one_run(
  run_number,
  files_folder,
  run_name,
  n_setups,
  n_runs_per_setup,
  date
)
}
\arguments{
\item{date}{}
}
\description{
Title
}
