% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_covariate.R
\name{add_covariate}
\alias{add_covariate}
\title{add_covariate}
\usage{
add_covariate(covariate_list, year)
}
\arguments{
\item{year}{}
}
\value{

}
\description{
adds the covariates based on the covariate_list and year vector
}
