% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture_error_measures_one_batch.R
\name{capture_error_measures_one_batch}
\alias{capture_error_measures_one_batch}
\title{Title}
\usage{
capture_error_measures_one_batch(
  location,
  format,
  array_name,
  date,
  i,
  batch_size,
  intercepts,
  trends,
  sigma,
  pi_int,
  pi_trend,
  sigma_tolerance = c(0.05, 100),
  pi_tolerance = c(0.05, 0.95),
  intercepts_tolerance = 100,
  trends_tolerance = 100,
  number_of_batches
)
}
\arguments{
\item{trends_tolerance}{}
}
\description{
Title
}
