% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture_error_measures_one_run.R
\name{capture_error_measures_one_run}
\alias{capture_error_measures_one_run}
\title{Title}
\usage{
capture_error_measures_one_run(results, settings)
}
\arguments{
\item{settings}{}
}
\description{
Title
}
