% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censor_values.R
\name{censor_values}
\alias{censor_values}
\title{censor_values}
\usage{
censor_values(simulated_obs, scale = "log")
}
\arguments{
\item{scale}{}
}
\description{
adds new columns with left and right censoring indicators, setting up the creation of a Surv object with interval2 format
}
