% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censor_values.R
\name{censor_values}
\alias{censor_values}
\title{censor_values}
\usage{
censor_values(df, MIC_breakpoint, low_log_con = -4, high_log_con = 4)
}
\arguments{
\item{df}{must contain a column named "observed value" which contains the uncensored simulated values}

\item{MIC_breakpoint}{is the highest MIC that is still S (the MIC where susceptible is defined as ≥ MIC), this value is on the MIC scale (we take log2 of MIC_breakpoint before applying it to the continuous values that are simulated on the log2 scale)}
}
\value{

}
\description{
adds new columns with left and right censoring indicators, setting up the creation of a Surv object with interval2 format
}
