% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censor_values.R
\name{censor_values}
\alias{censor_values}
\title{censor_values}
\usage{
censor_values(
  observed_value,
  low_con = 2^-4,
  high_con = 2^4,
  tested_concentrations = log2(low_con):log2(high_con),
  output_scale = "concentration"
)
}
\arguments{
\item{tested_concentrations}{}
}
\value{

}
\description{
adds new columns with left and right censoring indicators, setting up the creation of a Surv object with interval2 format
}
