% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component_mean.R
\name{component_mean}
\alias{component_mean}
\title{component_mean}
\usage{
component_mean(
  n = 100,
  t_dist = t_dist1,
  pi = pi1,
  mean_func = mean_func1,
  complist = complist1
)
}
\arguments{
\item{complist}{}
}
\value{

}
\description{
component_mean
}
