% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component_mean.R
\name{component_mean}
\alias{component_mean}
\title{component_mean}
\usage{
<<<<<<< HEAD
component_mean(
  n = 100,
  t_dist = function(n) {     runif(n, min = 0, max = 1) },
  pi = function(t) {     z <- 0.5 + 0.2 * t     c(z, 1 - z) },
  complist = list(f1 = function(t) {     3 + t + 2 * t^2 - sqrt(t) }, f2 = function(t)
    {     3 * t })
)
=======
component_mean(n = 100, t_dist = t_dist1, pi = pi1, complist = complist1)
>>>>>>> f9e90138c01c250990b562d0793552f368a42331
}
\arguments{
\item{complist}{}
}
\description{
component_mean
}
