% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_epsilon.R
\name{draw_epsilon}
\alias{draw_epsilon}
\title{draw_epsilon}
\usage{
draw_epsilon(
  n = 100,
<<<<<<< HEAD
  t_dist = function(n) {     runif(n, min = 0, max = 1) },
  pi = function(t) {     z <- 0.5 + 0.2 * t     c(z, 1 - z) },
  complist = list(`1` = function(t) {     3 + t + 2 * t^2 - sqrt(t) }, `2` =
    function(t) {     3 * t }),
=======
  t_dist = t_dist1,
  pi = pi1,
  complist = list(`1` = function(t) {
     3 + t + 2 * t^2 - sqrt(t)
 }, `2` =
    function(t) {
     3 * t
 }),
>>>>>>> f9e90138c01c250990b562d0793552f368a42331
  sd_vector = c(group1 = 1, group2 = 2)
)
}
\arguments{
\item{sd_vector}{}
}
\description{
draw_epsilon
}
