% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_epsilon.R
\name{draw_epsilon}
\alias{draw_epsilon}
\title{draw_epsilon}
\usage{
draw_epsilon(
  n = 100,
  t_dist = t_dist1,
  pi = pi1,
  mean_func = mean_func1,
  complist = complist1,
  sd_vector = c(`1` = 1, `2` = 2)
)
}
\arguments{
\item{sd_vector}{}
}
\value{

}
\description{
draw_epsilon
}
