% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_algorithm_reduced.R
\name{EM_algorithm_reduced}
\alias{EM_algorithm_reduced}
\title{Title}
\usage{
EM_algorithm_reduced(
  fixed_side = "RC",
  extra_row = FALSE,
  visible_data,
  model = "surv",
  mu_formula = Surv(time = left_bound, time2 = right_bound, type = "interval2") ~
    pspline(t, df = 0, caic = TRUE),
  pi_formula = c == "2" ~ s(t),
  max_it = 3000,
  ncomp = 2,
  tol_ll = 1e-06,
  browse_at_end = FALSE,
  browse_each_step = FALSE,
  plot_visuals = FALSE,
  prior_step_plot = FALSE,
  pause_on_likelihood_drop = FALSE,
  pi_link = "logit",
  verbose = 3,
  model_coefficient_tolerance = 1e-05,
  maxiter_survreg = 30,
  initial_weighting = 1,
  sd_initial = 0.2,
  stop_on_likelihood_drop = FALSE,
  non_linear_term = "t",
  covariates = NULL,
  scale = NULL
)
}
\arguments{
\item{covariates}{}
}
\description{
Title
}
