% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_measures_one_batch.R
\name{error_measures_one_batch}
\alias{error_measures_one_batch}
\title{error_measures_one_batch}
\usage{
error_measures_one_batch(
  location,
  format,
  array_name,
  date,
  i,
  intercepts,
  trends,
  sigma,
  pi,
  error_threshold = 10
)
}
\arguments{
\item{error_threshold}{}
}
\description{
error_measures_one_batch
}
