% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_measures_one_run_both_directions.R
\name{error_measures_one_run_both_directions}
\alias{error_measures_one_run_both_directions}
\title{Title}
\usage{
error_measures_one_run_both_directions(
  individual_run,
  intercepts,
  trends,
  sigma,
  pi,
  sigma_tolerance = 100,
  pi_tolerance = 100,
  intercepts_tolerance = 100,
  trends_tolerance = 100
)
}
\arguments{
\item{trends_tolerance}{}
}
\description{
Title
}
