% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_EM.R
\name{fit_EM}
\alias{fit_EM}
\title{Title}
\usage{
fit_EM(
  model = "surv",
  approach = "full",
  pre_set_degrees = NULL,
  max_degree = 8,
  degree_sets = "matched",
  visible_data,
  nfolds = 10,
  non_linear_term = "t",
  covariates = NULL,
  pi_formula = c == "2" ~ s(t),
  fixed_side = NULL,
  extra_row = FALSE,
  max_it = 3000,
  ncomp = 2,
  tol_ll = 1e-06,
  pi_link = "logit",
  verbose = 3,
  model_coefficient_tolerance = 1e-05,
  maxiter_survreg = 30,
  initial_weighting = 3,
  sd_initial = 0.2,
  scale = NULL,
  reruns_allowed = 3
)
}
\arguments{
\item{sd_initial}{}
}
\description{
Title
}
