% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lr.R
\name{fit_lr}
\alias{fit_lr}
\title{fit_lr}
\usage{
fit_lr(
  df,
  time = "t",
  covariate_names,
  left_bound,
  right_bound,
  MIC_breakpoint,
  summary = FALSE
)
}
\arguments{
\item{df}{data frame of year, observed values (colname is observed_value), and covariates}

\item{MIC_breakpoint}{on MIC scale, highest value that is still S (so if the guide says all organisms with an MIC ≤ 2  are S and ≥ 4 are R, the breakpoint is 2)}

\item{summary}{whether the output is the summary.glm  (TRUE) or the glm object}
}
\value{

}
\description{
fit_lr
}
