% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{Title}
\usage{
fit_model(
  visible_data,
  formula = Surv(time = left_bound, time2 = right_bound, type = "interval2") ~ 0 + c +
    strata(c) + t:c,
  max_it = 3000,
  ncomp = 2,
  tol_ll = 1e-06,
  browse_at_end = FALSE,
  browse_each_step = FALSE,
  plot_visuals = FALSE,
  verbose = 3,
  low_con = 2^-3,
  high_con = 2^3,
  maxiter_survreg = 30
)
}
\arguments{
\item{maxiter_survreg}{}
}
\description{
Title
}
