% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model_pi.R
\name{fit_model_pi}
\alias{fit_model_pi}
\title{fit_model_pi}
\usage{
fit_model_pi(
  visible_data,
  formula = Surv(time = left_bound, time2 = right_bound, type = "interval2") ~ pspline(t,
    df = 0, calc = TRUE),
  formula2 = c == "2" ~ s(t),
  max_it = 3000,
  ncomp = 2,
  tol_ll = 1e-06,
  browse_at_end = FALSE,
  browse_each_step = TRUE,
  plot_visuals = TRUE,
  prior_step_plot = TRUE,
  pi_link = "logit",
  verbose = 3,
  maxiter_survreg = 30,
  initial_weighting = 1
)
}
\arguments{
\item{initial_weighting}{}
}
\description{
fit_model_pi
}
