% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_spaft.R
\name{fit_spaft}
\alias{fit_spaft}
\title{fit_spaft}
\usage{
fit_spaft(
  df,
  time = "t",
  covariate_names,
  left_bound,
  right_bound,
  summary = FALSE
)
}
\arguments{
\item{covariate_names}{a vector with all the text names of the covariates (used to build a formula)}

\item{summary}{}
}
\description{
apply semiparametric accelerated failure time model using smoothSurv
}
