% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_surv_EM.R
\name{fit_surv_EM}
\alias{fit_surv_EM}
\title{Title}
\usage{
fit_surv_EM(
  pre_set_degrees = NULL,
  max_degree,
  degree_sets = "matched",
  visible_data,
  nfolds,
  non_linear_term = "t",
  covariates = NULL,
  pi_formula = c == "2" ~ s(t),
  max_it = 3000,
  ncomp = 2,
  tol_ll = 1e-06,
  pi_link = "logit",
  verbose = 3,
  model_coefficient_tolerance = 1e-05,
  initial_weighting = 9,
  sd_initial = 0.2
)
}
\arguments{
\item{sd_initial}{}
}
\description{
Title
}
