% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_sim_in_1_function.R
\name{full_sim_in_1_function}
\alias{full_sim_in_1_function}
\title{full_sim_in_1_function}
\usage{
full_sim_in_1_function(
  i,
  n = 150,
  t_dist = function(n) {
     runif(n, min = 0, max = 5)
 },
  pi = function(t) {
     z <- 0.6
     c(`1` = z, `2` = 1 - z)
 },
  `E[X|T,C]` = function(t, c) {
     case_when(c == "1" ~ -2 - 0.1 * t, c == "2" ~ 2 +
    0.2 * t, TRUE ~ NaN)
 },
  sd_vector = c(`1` = 1, `2` = 2),
  covariate_list = NULL,
  covariate_effect_vector = c(0),
  covariate_names = NULL,
  conc_limits_table = NULL,
  low_con = -4,
  high_con = 4,
  scale = "log",
  formula = Surv(time = left_bound, time2 = right_bound, type = "interval2") ~ 0 + c +
    strata(c) + t:c,
  formula2 = c == "2" ~ t,
  max_it = 3000,
  ncomp = 2,
  tol_ll = 1e-06,
  maxiter_survreg = 30,
  pi_function = TRUE,
  pi_link = "logit",
  verbose = 3,
  allow_safety = TRUE,
  cutoff = 0.9,
  fms_only = FALSE,
  initial_weighting = 1,
  ...
)
}
\arguments{
\item{cutoff}{}
}
\description{
full_sim_in_1_function
}
