% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_mics_with_metadata.R
\name{import_mics_with_metadata}
\alias{import_mics_with_metadata}
\title{Title}
\usage{
import_mics_with_metadata(
  data,
  mic_column,
  metadata_columns = NULL,
  code_column = NULL,
  combination_agent = NULL,
  log_reg_value = FALSE,
  scale = "log",
  round = FALSE
)
}
\arguments{
\item{log_reg_value}{}
}
\description{
Title
}
