% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_mics_with_metdata.R
\name{import_mics_with_metadata}
\alias{import_mics_with_metadata}
\title{Title}
\usage{
import_mics_with_metadata(
  mic_column,
  metadata_cols = NULL,
  code_column = NULL,
  combination_agent = FALSE,
  log_reg_value = FALSE
)
}
\arguments{
\item{log_reg_value}{}
}
\description{
Title
}
