% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_reg.R
\name{log_reg}
\alias{log_reg}
\title{Title}
\usage{
log_reg(
  data,
  split_by = "ecoff",
  data_type,
  drug,
  date_col,
  date_type,
  first_year,
  s_breakpoint,
  r_breakpoint,
  ecoff,
  visual_split,
  k = NULL
)
}
\arguments{
\item{data}{either possible_data or a set of mic data ready you want to run import_mics on}

\item{data_type}{use either "possible_data" if passing in a possible_data object or "import" if you want to import mics and run logistic regression}

\item{drug}{NULL if using possible_data, if importing data should be the name of the column of the mics or a vector of the mic column and the sign colomn}

\item{date_col}{string, what is the name of the column in the data that corresponds to time of sampling}

\item{date_type}{string, either "decimal", "date", or "year" use decimal if using t from possible data, date or year if importing mic data and the date column is a date or just a year respectively}

\item{first_year}{NULL if date_type is "decimal", otherwise a numeric year or decimal year value if using "year" or "date" for date_type respectively}

\item{s_breakpoint}{string, the breakpoint on the MIC scale for what constitutes a susceptible isolate, e.g. <=8 (ug/mL, do not include units)}

\item{r_breakpoint}{string, the breakpoint on the MIC scale for what constitutes a resistant isolate, e.g. >=128 (ug/mL, do not include units)}

\item{ecoff}{string or numeric, see plot_fm()}

\item{visual_split}{string or numeric, see plot_fm()}

\item{k}{variable passed into logistic regression GAM}
}
\description{
Title
}
