% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mic.sim-package.R
\docType{package}
\name{mic.sim-package}
\alias{mic.sim}
\alias{mic.sim-package}
\title{mic.sim: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ajmichaelucd/mic.sim}
  \item \url{https://ajmichaelucd.github.io/mic.sim/}
  \item Report bugs at \url{https://github.com/ajmichaelucd/mic.sim/issues}
}

}
\keyword{internal}
