% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mic.sim-package.R
\docType{package}
\name{mic.sim-package}
\alias{mic.sim}
\alias{mic.sim-package}
\title{mic.sim: Simulating and Fitting Models to MIC Data}
\description{
mic.sim includes functions for analyzing minimum inhibitory concentration data, including functions for importing or simulating data, viewing the distribution, model fitting, and model visualization. The model fitting uses a latent Gaussian mixture model framework to analyze trends in the mixture weights and component means on the log2 MIC scale to attempt to distinguish and model the wild type and non-wild type subpopulations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ajmichaelucd/mic.sim}
  \item \url{https://ajmichaelucd.github.io/mic.sim/}
  \item Report bugs at \url{https://github.com/ajmichaelucd/mic.sim/issues}
}

}
\keyword{internal}
