% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fm.R
\name{plot_fm}
\alias{plot_fm}
\title{Title}
\usage{
plot_fm(
  output,
  title,
  add_log_reg = FALSE,
  s_breakpoint = NA,
  r_breakpoint = NA,
  use_prior_step = FALSE,
  range_zoom = FALSE,
  plot_range = NULL,
  start_date = 2007
)
}
\arguments{
\item{r_breakpoint}{}
}
\description{
Title
}
