% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mean.R
\name{plot_mean}
\alias{plot_mean}
\title{Title}
\usage{
plot_mean(
  output,
  df,
  results,
  start_date,
  fitted_comp,
  title,
  plot_min,
  plot_max,
  ecoff,
  s_breakpoint,
  r_breakpoint,
  visual_split,
  x_axis_t_breaks,
  assumed_components,
  n_fitted_components
)
}
\arguments{
\item{n_fitted_components}{}
}
\description{
Title
}
\keyword{internal}
