% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pi.R
\name{plot_pi}
\alias{plot_pi}
\title{Title}
\usage{
plot_pi(
  output,
  df,
  start_date,
  add_log_reg,
  ecoff,
  s_breakpoint,
  r_breakpoint,
  visual_split,
  skip,
  x_axis_t_breaks
)
}
\arguments{
\item{x_axis_t_breaks}{}
}
\description{
Title
}
\keyword{internal}
