% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_sim_data_for_em.R
\name{prep_sim_data_for_em}
\alias{prep_sim_data_for_em}
\title{Intermediate function to prepare simulated data for use in EM algorithm}
\usage{
prep_sim_data_for_em(
  data.sim = simulate_mics(),
  left_bound_name = "left_bound",
  right_bound_name = "right_bound",
  time = "t",
  covariate_names = NULL,
  scale = NULL,
  observed_value_choice = FALSE,
  observed_value_name = "observed_value",
  low_con_name = "low_con",
  high_con_name = "high_con"
)
}
\arguments{
\item{high_con_name}{}
}
\description{
Intermediate function to prepare simulated data for use in EM algorithm
}
