% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_sim_data_for_em.R
\name{prep_sim_data_for_em}
\alias{prep_sim_data_for_em}
\title{Intermediate function to prepare simulated data for use in EM algorithm}
\usage{
prep_sim_data_for_em(
  data.sim,
  left_bound_name = "left_bound",
  right_bound_name = "right_bound",
  time = "t",
  covariate_names = NULL,
  scale = NULL
)
}
\arguments{
\item{scale}{}
}
\description{
Intermediate function to prepare simulated data for use in EM algorithm
}
