% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview_data.R
\name{preview_data}
\alias{preview_data}
\title{Plot Data}
\usage{
preview_data(data, title = "", y_min = NULL, y_max = NULL)
}
\arguments{
\item{data}{Tibble or data frame from simulate_mics or import_mics_with_metadata}

\item{title}{Title of plot}

\item{y_min}{Minimum value for plot, some extra space will be added below so the lowest tested concentration is a reasonable value}

\item{y_max}{Maximum value for plot, some extra space will be added below so the highest tested concentration is a reasonable value}
}
\description{
Produces a basic plot of data generated from simulate_mics or import_mics_with_metadata
}
