% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recreate_data_set.R
\name{recreate_data_set}
\alias{recreate_data_set}
\title{recreate_data_set}
\usage{
recreate_data_set(
  i,
  n,
  intercepts,
  trends,
  sigma,
  pi,
  nyears,
  low_con = 2^-3,
  high_con = 2^2,
  scale = "log"
)
}
\arguments{
\item{scale}{}
}
\description{
recreate_data_set
}
