% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{sim_data}
\alias{sim_data}
\title{sim_data}
\usage{
sim_data(nyears, sample_size_dist, norm_mean, norm_sd, unif_min, unif_max)
}
\arguments{
\item{unif_max}{}
}
\value{

}
\description{
Function to draw values for each year
}
