% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data_and_add_covariates.R
\name{sim_data_and_add_covariates}
\alias{sim_data_and_add_covariates}
\title{Title}
\usage{
sim_data_and_add_covariates(
  nyears = 5,
  sample_size_dist = "normal",
  norm_mean = 100,
  norm_sd = 10,
  unif_min,
  unif_max,
  covariate_list,
  initial_lower_mean = -1,
  initial_upper_mean = 1,
  initial_lower_std_dev = 1,
  initial_upper_std_dev = 1,
  initial_lower_slp = 0,
  initial_upper_slp = 0,
  initial_lower_slp_qd = 0,
  initial_upper_slp_qd = 0,
  initial_lower_lambda = 0.5,
  initial_upper_lambda = 0.5,
  initial_lower_slp_lambda = 0,
  initial_upper_slp_lambda = 0
)
}
\arguments{
\item{initial_upper_slp_lambda}{}
}
\value{

}
\description{
Title
}
