% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_mics.R
\name{simulate_mics}
\alias{simulate_mics}
\title{simulate mics}
\usage{
simulate_mics(
  year,
  sd_1 = 1,
  sd_2 = 1,
  mean_1_trend = 0,
  mean_2_trend = 0,
  mean_1_intercept = -1,
  mean_2_intercept = 1,
  pi_1_trend = 0,
  pi_1_intercept = 0.5,
  covariate_list,
  covariate_effect_vector
)
}
\arguments{
\item{covariate_effect_vector}{}
}
\value{

}
\description{
wraps together find_epsilon, add_covariate, covariate_effect_total to produce a tibble with all those outputs included
}
