% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_all_formulas.R
\name{write_all_formulas}
\alias{write_all_formulas}
\title{Title}
\usage{
write_all_formulas(non_linear_term, degrees, covariates, model)
}
\arguments{
\item{non_linear_term}{String, non-linear term to be included in the model. Variable in the pspline term in the pspline model or in the polynomial term in the polynomial model.}

\item{covariates}{String, covariates to be included in mu model aside from the non-linear term.}

\item{model}{}
}
\description{
Title
}
